MultiLineTextComponent.svelte resizes based on text to be shown on mount and on refresh.
