+++ packages/core-svelte/src/lib/components/FreonComponent.svelte
@@ -53,7 +53,7 @@
     }
 
     const onKeyDown = (event: KeyboardEvent) => {
-        console.log(
+        LOGGER.log(
             'FreonComponent onKeyDown: ' +
                 event.key +
                 ' ctrl: ' +
