+++ packages/core-svelte/src/lib/components/KeyBindings.md
@@ -1,0 +1,11 @@
+# Keyboard Entries and their Function
+
+[//]: # (TODO make a list for the help page in the webapp)
+
+In Multiline text:
+
+- Ctrl-enter: go to next editable field
+
+In single line text:
+
+- Enter: go to next editable field
