+++ packages/samples/Example/src/defs/LanguageDefinition-default.edit
@@ -7,7 +7,7 @@ global {
 
 
 Attribute {
-    [ ${name} : ${declaredType} ${extra multiline}]
+    [ ${name} : ${declaredType} ]
 }
 Entity {
     [
